% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_get_pm.R
\name{create_get_pm_fun}
\alias{create_get_pm_fun}
\title{Make \code{get_pm} function}
\usage{
create_get_pm_fun(f)
}
\arguments{
\item{f}{fun. This function cannot take a chr vector of length 1.}
}
\value{
fun. A function that take a chr vector (length >= 1) and apply the function \code{f}.
}
\description{
The functions \verb{get_nodesdf_*} and \code{get_nodedf_} work with a chr vector of length 1.
\code{create_get_pm_fun} can make that those functions can operate with vector of length > 1.
The \code{pm} means PubMed.
}
\examples{
get_pm_authors <- create_get_pm_fun(get_authors)
}
