% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_abstracts.R
\name{get_abstract}
\alias{get_abstract}
\title{Extract the abstracts with associated data from one PubMed record}
\usage{
get_abstract(doc)
}
\arguments{
\item{doc}{str. An PubMed xml record of one record.}
}
\value{
df. A table with the text, label and Nlm category
}
\description{
Extract the abstracts with associated data from one PubMed record
}
\examples{

system.file("extdata", "one_record_male_cancer01.txt", package = "fastpubmedxmlparser") |>
  readPubmedFiles() |>
  get_abstract()
}
