% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_authors.R
\name{get_authors}
\alias{get_authors}
\title{Extract the author data from a unique PubMed xml record character}
\usage{
get_authors(doc)
}
\arguments{
\item{doc}{str. An PubMed xml record of one record.}
}
\value{
df. Author data included: laste name, fore name, initial and affiliation
}
\description{
Extract the author data from a unique PubMed xml record character
}
\examples{

system.file("extdata", "one_record_male_cancer01.txt", package = "fastpubmedxmlparser") |>
  readPubmedFiles() |>
  get_authors()
}
