% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodedf.R
\name{get_nodedf_journal}
\alias{get_nodedf_journal}
\alias{get_nodedf_country}
\alias{get_nodedf_language}
\alias{get_nodedf_title}
\alias{get_nodedf_year}
\title{Get a dataframe from a single node for a unique PubMed record.}
\usage{
get_nodedf_journal(doc)

get_nodedf_country(doc)

get_nodedf_language(doc)

get_nodedf_title(doc)

get_nodedf_year(doc)
}
\arguments{
\item{doc}{chr. A PubMed record in xml format}
}
\value{
\itemize{
\item df. The first column is the node name.
\item \code{NULL} in case the node has no name.
}
}
\description{
The functions \verb{get_nodedf_*} all are derived functions from \code{create_node_to_df}.

These find all the first nodes for a specific \code{xpath} and return the text of such node.

It only works in a vector of length 1.
}
\examples{
system.file("extdata", "one_record_male_cancer01.txt", package = "fastpubmedxmlparser") |>
  readPubmedFiles() |>
  get_nodedf_journal()
}
