% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nodedf.R
\name{get_nodedf_keywords}
\alias{get_nodedf_keywords}
\alias{get_nodedf_mesh_description}
\alias{get_nodedf_mesh_qualifier}
\alias{get_nodedf_article_ids}
\alias{get_nodedf_publication_type}
\title{Get a dataframe from a list of nodes for a unique PubMed record.}
\usage{
get_nodedf_keywords(doc)

get_nodedf_mesh_description(doc)

get_nodedf_mesh_qualifier(doc)

get_nodedf_article_ids(doc)

get_nodedf_publication_type(doc)
}
\arguments{
\item{doc}{chr. A PubMed record in xml format}
}
\value{
\itemize{
\item df. The first column is the node name, the rest has the value of each attribute.
\item \code{NULL} in case there is no nodes.
}
}
\description{
The functions \verb{get_nodedf_*} all are derived functions from \code{create_node_to_df}.

These find all the nodes for a specific \code{xpath}, and they also use the \code{attr} of those nodes.

It only works in a vector of length 1.
}
\examples{
system.file("extdata", "one_record_male_cancer01.txt", package = "fastpubmedxmlparser") |>
  readPubmedFiles() |>
  get_nodedf_mesh_description()
}
