% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pm.R
\name{get_pm_authors}
\alias{get_pm_authors}
\alias{get_pm_abstract}
\title{Extract metadata from a record list}
\usage{
get_pm_authors(str_list)

get_pm_abstract(str_list)
}
\arguments{
\item{str_list}{chr. A chr vector with multiple vect. Usually extracted from \code{readPubmedFiles} function.}
}
\value{
A list of the same length of str_list paragram
}
\description{
Extract metadata from a record list
}
\examples{
system.file("extdata", "one_record_male_cancer01.txt", package = "fastpubmedxmlparser") |>
  get_pm_abstract() |>
  pubmed_list2df()

}
