% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_list2df.R
\name{pubmed_list2df}
\alias{pubmed_list2df}
\title{Make a unique dataframe from a list of \code{dataframes} of the same PubMed data}
\usage{
pubmed_list2df(dfs)
}
\arguments{
\item{dfs}{ls. A list, typicall return by one \verb{get_pm*} function.}
}
\value{
df. A table with the same columns
}
\description{
Make a unique dataframe from a list of \code{dataframes} of the same PubMed data
}
\examples{
system.file("extdata", "pubmed_male_cancer_10.txt", package = "fastpubmedxmlparser") |>
  get_pm_authors() |>
  pubmed_list2df()
}
