% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPubmedFiles.R
\name{readPubmedFiles}
\alias{readPubmedFiles}
\title{#' Read a XML PubMed file into character strings}
\usage{
readPubmedFiles(file_path)
}
\arguments{
\item{file_path}{chr vector. The path for the PubMed record file.}
}
\value{
A character vector, with each PubMed record represented as a string.
}
\description{
This function reads an XML PubMed file and split it into strings
by the PubmedArticle tag.
}
\examples{
# From local file ----------------------------------------------------------------
if (FALSE){

library(fastpubmedxmlparser)

#Suppose you have a PubMed record in the pubmedData directory.
records <- 'pubmedData/your-pubmed-record.txt' |>
  readPubmedFiles()

}

# Using the file include in this pkg-----------------------------------------------
records <- system.file("extdata", "maleCancer.txt", package = "fastpubmedxmlparser") |>
readPubmedFiles() |>
str()

}
\seealso{
\link{get_pm_authors}, \link{get_pm_abstract}, and all the \verb{get_pm_*}
functions for futher analysis.
}
