% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_get_pm.R
\name{create_get_pm_fun}
\alias{create_get_pm_fun}
\title{Make \code{get_pm} function}
\usage{
create_get_pm_fun(f)
}
\arguments{
\item{f}{fun. A function that take one chr}
}
\value{
fun. A function that take a chr vector with length of one or more and apply the function \code{f}.
}
\description{
Make \code{get_pm} function
}
\examples{
get_pm_authors <- create_get_pm_fun(get_authors)
}
