% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_node_to_df.R
\name{create_node_to_df}
\alias{create_node_to_df}
\title{Dynamic generation of a single XML node to dataframe Functions}
\usage{
create_node_to_df(xpath)
}
\arguments{
\item{xpath}{chr. A path of the node}
}
\value{
fun. For applications see a \verb{get_nodedf_*} function
}
\description{
This function is similar to \code{create_node_to_df} but it will use the first node.
The returning function, internally calls \code{xml2::xml_find_first}.
}
\examples{
#This function is used to generate the `get_nodedf_*` functions
get_nodedf_title <- create_node_to_df('.//ArticleTitle')
}
