% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pm.R
\name{get_pm_authors}
\alias{get_pm_authors}
\alias{get_pm_abstract}
\alias{get_pm_keywords}
\alias{get_pm_mesh_des}
\alias{get_pm_mesh_qua}
\alias{get_pm_paper_ids}
\alias{get_pm_pub_type}
\alias{get_pm_journal}
\alias{get_pm_country}
\alias{get_pm_language}
\alias{get_pm_title}
\alias{get_pm_year}
\title{Extract metadata from a record list}
\usage{
get_pm_authors(str_list)

get_pm_abstract(str_list)

get_pm_keywords(str_list)

get_pm_mesh_des(str_list)

get_pm_mesh_qua(str_list)

get_pm_paper_ids(str_list)

get_pm_pub_type(str_list)

get_pm_journal(str_list)

get_pm_country(str_list)

get_pm_language(str_list)

get_pm_title(str_list)

get_pm_year(str_list)
}
\arguments{
\item{str_list}{chr. A character vector with multiple elements. Usually
extracted from the \link{readPubmedFiles} function.}
}
\value{
A list of the same length of str_list parameter. Each element of the
list is a dataframe. The first column will contain the text of the nodeset
in the xml record. The subsequent columns names will be named with the
attributes of the nodeset, and the rows will contain the attribute value.
If no nodeset was found for a record, a NULL value will be return. Each
element is named with the PMID of the record
}
\description{
Extract metadata from a record list
}
\examples{
system.file("extdata", "one_record_male_cancer01.txt", package = "fastpubmedxmlparser") |>
  readPubmedFiles() |>
  get_pm_abstract() |>
  str()

system.file("extdata", "pubmed_male_cancer_10.txt", package = "fastpubmedxmlparser") |>
  readPubmedFiles() |>
  get_pm_pub_type() |>
  str()

}
