% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmed_list2df.R
\name{pubmed_list2df}
\alias{pubmed_list2df}
\title{Create a single dataframe from a list of dataframes containing the same PubMed data}
\usage{
pubmed_list2df(dfs)
}
\arguments{
\item{dfs}{list. A list of dataframes, typicall return by one of the \verb{get_pm*} functions.}
}
\value{
df. A dataframe that combines the columns from the input list of dataframes.
}
\description{
Create a single dataframe from a list of dataframes containing the same PubMed data
}
\examples{
system.file("extdata", "pubmed_male_cancer_10.txt", package = "fastpubmedxmlparser") |>
  readPubmedFiles() |>
  get_pm_keywords() |>
  pubmed_list2df()

}
