% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPubmedFiles.R
\name{readPubmedFiles}
\alias{readPubmedFiles}
\title{Read a xml pubmed file into chr strings}
\usage{
readPubmedFiles(file_path)
}
\arguments{
\item{file_path}{chr vect. The path for the PubMed record.}
}
\value{
chr vect. With each record as a character vector
}
\description{
Split the xml file into strings, according to the number of PubmedArticles that the XML file contains
}
\examples{
# From local file ----------------------------------------------------------------
if (FALSE){

library(fastpubmedxmlparser)

#Suppose you have a PubMed record in the pubmedData directory.
records <- 'pubmedData/your-pubmed-record.txt' |>
  readPubmedFiles()

}

# Using the file include in this pkg-----------------------------------------------
records <- system.file("extdata", "maleCancer.txt", package = "fastpubmedxmlparser") |>
readPubmedFiles()

records |> str()


}
